File_Upload/requirements.txt
streamlit
pandas
python-dotenv
langchain
langchain-core
langchain-community
langchain-groq
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
openpyxl